requirements.txt
aniso8601==9.0.1
bidict==0.23.1
blinker==1.8.2
boto3==1.35.28
botocore==1.35.28
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-RESTful==0.3.10
gevent==24.2.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
packaging==24.1
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
simple-websocket==1.0.0
six==1.16.0
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
zope.event==5.0
zope.interface==7.0.3
